% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_bliss}
\alias{calc_bliss}
\title{Calculate Bliss scores of combinations for a single experiment}
\usage{
calc_bliss(tb)
}
\arguments{
\item{tb}{A tibble of dose effect data.

Columns: cell_line (string), drug1 (string), drug2 (string), dose1 (double), dose2 (double), effect (double)}
}
\value{
\item{A tibble}{Columns: cell_line, drug1, drug2, dose1, dose2, effect1 (mean), effect2 (mean), effect12 (mean), bliss}
}
\description{
Bliss = mean effect of comb drugs  / (mean effect of drug1 * mean effect of drug2)
}
\examples{
library(decurve)
tb <- calc_bliss(tb)
}
