% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decurve-package.R
\docType{package}
\name{decurve-package}
\alias{decurve-package}
\alias{decurve}
\title{decurve}
\description{
Dose Effect Curve
}
\details{
decurve is used to tidy and plot dose effect curves
}
\keyword{package}
