% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dose_effect}
\alias{plot_dose_effect}
\title{Dose effect curve}
\usage{
plot_dose_effect(tb)
}
\arguments{
\item{tb}{A tibble of dose effect data.

Columns: cell_line (string), drug1 (string), drug2 (string), dose1 (double), dose2 (double), effect (double)}
}
\value{
\item{A ggplot object}{returned ggplot object by ggline}
}
\description{
Plot dose effect curve of a single experiment
}
\examples{
library(decurve)
p <- plot_dose_effect(tb)
print(p)
}
