% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dose_GR}
\alias{plot_dose_GR}
\title{Dose GR curve}
\usage{
plot_dose_GR(tb)
}
\arguments{
\item{tb}{A tibble of dose GR data.

Columns: cell_line (string), drug1 (string), drug2 (string), dose1 (double), dose2 (double), GR (double)}
}
\value{
\item{A ggplot object}{returned ggplot object by ggline}
}
\description{
Plot dose GR curve of a single experiment
}
\examples{
library(decurve)
p <- plot_dose_GR(tb)
print(p)
}
